+++ out/language/main.cjs
@@ -35019,7 +35019,7 @@ function getReferenceablesInStatement(statement, output) {
     output.add(reference);
 }
 function getReferencablesInExpression(expression, output) {
-  if (expression == void 0)
+  if (expression === void 0)
     return;
   switch (expression.$type) {
     case "OrExpression":
@@ -35040,8 +35040,7 @@ function getReferencablesInExpression(expression, output) {
 }
 function getReferencablesInWhenCondition(condition) {
   let result = /* @__PURE__ */ new Set();
-  if (condition.expression != void 0)
-    getReferencablesInExpression(condition.expression, result);
+  getReferencablesInExpression(condition.expression, result);
   return result;
 }
 function getAllUsedConcerns(model) {
@@ -35069,7 +35068,7 @@ function getAllUsedReferenceables(model) {
         });
       });
     });
-    if (proposition.disable == void 0)
+    if (proposition.disable === void 0)
       return;
     proposition.disable.statements.forEach((disableStatement) => {
       getReferencablesInWhenCondition(disableStatement.condition).forEach((referenceable) => {
@@ -35099,6 +35098,9 @@ function registerValidationChecks2(services) {
     ],
     LaboratoryInformation: [
       validator.noDuplicateFieldsInLaboratoryInformation
+    ],
+    Statement: [
+      validator.statementReferencesValidValue
     ]
   };
   registry.register(checks, validator);
@@ -35210,6 +35212,31 @@ var JavaScriptPropositionalLaboratoryFormatValidator = class {
       accept("error", "Multiple authors for one laboratory are not allowed.", { node: information });
     if (information.versions.length > 1)
       accept("error", "Multiple versions for one laboratory are not allowed.", { node: information });
+  }
+  statementReferencesValidValue(statement, accept) {
+    if (statement === void 0)
+      return;
+    if (statement.value === void 0)
+      return;
+    if (statement.reference === void 0)
+      return;
+    if (statement.reference.ref === void 0)
+      return;
+    const referenceable = statement.reference.ref;
+    const value = statement.value;
+    if (referenceable.$type === "Condition") {
+      if (typeof value === "boolean")
+        return;
+      accept("error", "Stated value is not a valid value of the referenced object.", { node: statement, property: "value" });
+      return;
+    }
+    const proposition = referenceable;
+    if (proposition.valueClauses === void 0)
+      return;
+    const foundValue = proposition.valueClauses.map((clause) => clause.value).find((defined) => defined === value);
+    if (foundValue !== void 0)
+      return;
+    accept("error", "Stated value is not a valid value of the referenced object.", { node: statement, property: "value" });
   }
 };